Introducing the "pull ups"
Just considered how pull ups are used to indicate certain infromation and how it is pulled from the bottom to the top. I don't know how that would look if applied but really looking forward to it. the life of an amateur developer. That is if you can call me that even